+++ .storybook/webpack.config.js
@@ -28,7 +28,6 @@ module.exports = ({ config }) => {
     schemas: path.resolve(basePath, 'src/schemas'),
     src: path.resolve(basePath, 'src'),
     'ipfs-api': 'ipfs-api/dist',
-    'ipfs-http-client': 'ipfs-http-client/dist'
   };
   return config;
 };