+++ src/components/Proposal/ProposalStatus.tsx
@@ -1,7 +1,7 @@
 import { IProposalStage, IProposalState } from "@daostack/client";
 import * as classNames from "classnames";
 import * as React from "react";
-import { proposalExpired, proposalFailed, proposalPassed } from "reducers/arcReducer";
+import { proposalExpired, proposalFailed, proposalPassed } from "lib/proposalHelpers";
 import * as css from "./ProposalStatus.scss";
 
 export default class ProposalStatus extends React.Component<IProps, null> {
@@ -54,7 +54,7 @@ export default class ProposalStatus extends React.Component<IProps, null> {
                     <div className={classNames({
                       [css.status]: true,
                       [css.pending]: true,
-                    })}><img src="/assets/images/Icon/pending.svg" />Pending</div> :
+                    })}><img src="/assets/images/Icon/pending.svg" />Pending Boosting</div> :
 
                     (proposalState.stage === IProposalStage.Boosted) ?
                       <div className={classNames({
@@ -66,7 +66,7 @@ export default class ProposalStatus extends React.Component<IProps, null> {
                         <div className={classNames({
                           [css.status]: true,
                           [css.quietEnding]: true,
-                        })}><img src="/assets/images/Icon/boosted.svg" />Quiet Ending</div> :
+                        })}><img src="/assets/images/Icon/boosted.svg" />Boosted (overtime)</div> :
 
                         <div className={classNames({
                           [css.status]: true,