+++ src/components/Scheme/SchemeProposalsPage.tsx
@@ -1,10 +1,10 @@
 import * as H from "history";
-import { Address, IDAOState, IProposalStage, ISchemeState, Proposal, Vote, Reward, Stake } from "@daostack/client";
+import { Address, IDAOState, IProposalStage, ISchemeState, Proposal, Vote, Reward, Scheme, Stake } from "@daostack/client";
 import { enableWalletProvider, getArc } from "arc";
 import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import gql from "graphql-tag";
-import { schemeName} from "lib/util";
+import { schemeName } from "lib/util";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 import * as InfiniteScroll from "react-infinite-scroll-component";
@@ -120,7 +120,7 @@ class SchemeProposalsPage extends React.Component<IProps, null> {
             <div className={css.proposalsHeader}>
               No upcoming proposals
             </div>
-            <p>You can be the first one to create a {scheme.name && scheme.name.replace(/([A-Z])/g, " $1") || scheme.address} proposal today! (:</p>
+            <p>You can be the first one to create a {schemeName(scheme)} proposal today! (:</p>
             <div className={css.cta}>
               <Link to={"/dao/" + daoState.address}>
                 <img className={css.relax} src="/assets/images/lt.svg"/> Back to schemes
@@ -157,8 +157,8 @@ class SchemeProposalsPage extends React.Component<IProps, null> {
 
             <div className={css.regularContainer}>
               <div className={css.proposalsHeader}>
-                <TrainingTooltip placement="bottom" overlay={"Pending proposals have reached the prediction score required for boosting and now must make it through the pending period without dipping below that threshold in order to be boosted."}>
-                  <span>Pending Proposals ({scheme.numberOfPreBoostedProposals})</span>
+                <TrainingTooltip placement="bottom" overlay={"Pending boosting proposals have reached the prediction score required for boosting and now must make it through the pending period without dipping below that threshold in order to be boosted."}>
+                  <span>Pending Boosting Proposals ({scheme.numberOfPreBoostedProposals})</span>
                 </TrainingTooltip>
                 {proposalsPreBoosted.length === 0
                   ?
@@ -187,6 +187,7 @@ class SchemeProposalsPage extends React.Component<IProps, null> {
               </div>
               <div className={css.proposalsContainer}>
                 <InfiniteScroll
+                  style={{overflow: "visible"}}
                   dataLength={proposalsQueued.length} //This is important field to render the next data
                   next={fetchMore}
                   hasMore={proposalsQueued.length < scheme.numberOfQueuedProposals}
@@ -253,6 +254,7 @@ const SubscribedSchemeProposalsPage = withSubscription<IProps, SubscriptionData>
         ${Vote.fragments.VoteFields}
         ${Stake.fragments.StakeFields}
         ${Reward.fragments.RewardFields}
+        ${Scheme.fragments.SchemeFields}
       `;
     } else {
       bigProposalQuery = gql`
@@ -270,6 +272,7 @@ const SubscribedSchemeProposalsPage = withSubscription<IProps, SubscriptionData>
           }
         }
         ${Proposal.fragments.ProposalFields}
+        ${Scheme.fragments.SchemeFields}
       `;
     }