+++ CHANGELOG.md
@@ -6,6 +6,63 @@
   - Features Added
   - Bugs Fixed
 
+## 2020-01-22
+
+  - TLDR 
+    - Many tweaks to the language throughout the app
+    - Create proposal urls can now be exported and shared
+    - Improvements in feed and feed representation
+    - Various UI improvements
+  
+  - Features Added
+    - Can click on whole proposal card to go to proposal details page
+    - Can click on whole proposal history row to go to proposal details page
+    - Can click on DAO member row to go to account profile page
+    - Improvements to language of disabled vote and stake buttons
+    - Improvements to language of proposal countdown to show what status it will change to
+    - Can now press enter to vote or stake from pre transaction modal, or ESC to close the modal
+    - Add follow buttons to feed items
+    - fix display for really small screens where the feed item timestamp would mess up layout of feed items
+    - Add follow button to DAO sidebar
+    - Add modal to ask about connecting to 3box
+    - And speed up 3box  use by storing in redux for later use
+    - Get logging out of 3box to work when pressing Logout or switching accounts
+    - Lessen box shadow on feed items
+    - Update banner text on not connected feed page
+    - Don't show "not following anything" while loading feed
+    - Load less events on initial page load
+    - Change event title text for proposal state changes
+    - Change "Quiet Ending" to "Boosted (overtime)"
+    - Always show Show proposal details link at the bottom of every proposal feed item
+    - Dont show 3box interstitial if 3box has cached their signature
+    - Change text of follow button while it is pending a change
+    - Fade out bottom of proposal description in feed
+    - Tabbing in create proposal form is now more usable
+    - Upgrade client and subgraph dependencies (which fixes various bugs)
+    - Correctly wrap new lines in proposal descriptions
+    - Dont show "Proposal" before every proposal title
+    - Proposal feed item: Fix fade out to not cover Show more details link
+    - Don't show expired proposal feed items when following a DAO
+     -Also correctly show proposal items as expired
+
+  - Bugs Fixed
+    - Fixed up meta tags for various pages and when sharing to Twitter and Facebook
+    - If proposal card action menu would go off screen to the right then it now appears to the left of the button
+    - Improve performance of 3box actions like follow items for feed.
+    - Add informational popup before any action that interacts with 3box.
+    - Add follow button to DAO sidebar
+    - Improve copy of home page when not logged in
+    - Improve and clean up header text for feed items
+    - Fix bug where tooltips in sidebar would cover other menu items. Move them to the right of the sidebar
+    - Log out of 3box when clicking the Log Out button
+    - Recognize latest ENS contract
+    - Remove duplicate privacy policy
+    - Fix reputation display for vote and stake events
+    - Fix display of enable predictions button in proposal action menu
+    - fix ability to execute proposals when DAO doesnt have funds to redeem for beneficiary.
+    - Fix display of external token rewards
+
+
 ## 2019-12-18
 
   - Features Added
@@ -17,9 +74,9 @@
 
 ## 2019-12-10
   - Features Added
-    - Use 3box profiles 
+    - Use 3box profiles
     - Create a Feed of events
-    - add new cookie policy and privacy policy 
+    - add new cookie policy and privacy policy
   - Bugs Fixed
     - do not error on bad token address
     - convert deprecated React methods
@@ -27,7 +84,7 @@
     - add subscriptions for votes, stakes and rewards on the proposal page
     - ignore externalTokenReward when address is null
     - change old dai to sai, add new dai
-    - better position notifications 
+    - better position notifications
     - shorten token dropdown menu
 
 ## 2019-12-03